PHOENIX-5616 reverting changes to fix qa build failure

+++ egov/egov-stms/src/main/java/org/egov/stms/entity/es/SewerageIndex.java
@@ -116,10 +116,10 @@ public class SewerageIndex {
     @Field(type = FieldType.String, index = FieldIndex.not_analyzed)
     private String propertyType;
 
-    @Field(type = FieldType.Long)
+    @Field(type = FieldType.Integer)
     private Integer noOfClosets_residential;
 
-    @Field(type = FieldType.Long)
+    @Field(type = FieldType.Integer)
     private Integer noOfClosets_nonResidential;
 
     @Field(type = FieldType.String, index = FieldIndex.not_analyzed)
